@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Text;
@@ -7,7 +8,25 @@ namespace TrRebootTools.Shared.Util
 {
     public static class IoExtensions
     {
-        private static readonly byte[] TempBuffer = new byte[0x400];
+        private static readonly byte[] TempBuffer = new byte[0x2000];
+
+        public static float[] ReadFloatArray(this BinaryReader reader, int count)
+        {
+            float[] result = new float[count];
+            for (int i = 0; i < count; i++)
+            {
+                result[i] = reader.ReadSingle();
+            }
+            return result;
+        }
+
+        public static void WriteFloatArray(this BinaryWriter writer, IEnumerable<float> values)
+        {
+            foreach (float value in values)
+            {
+                writer.Write(value);
+            }
+        }
 
         public static string ReadZeroTerminatedString(this BinaryReader reader)
         {
@@ -28,6 +47,46 @@ namespace TrRebootTools.Shared.Util
             int length = Encoding.UTF8.GetBytes(text, 0, text.Length, TempBuffer, 0);
             TempBuffer[length++] = 0;
             writer.Write(TempBuffer, 0, length);
+        }
+
+        public static int Tell(this BinaryReader reader)
+        {
+            return (int)reader.BaseStream.Position;
+        }
+
+        public static int Tell(this BinaryWriter writer)
+        {
+            return (int)writer.BaseStream.Position;
+        }
+
+        public static void Seek(this BinaryReader reader, int position)
+        {
+            reader.BaseStream.Position = position;
+        }
+
+        public static void Seek(this BinaryWriter writer, int position)
+        {
+            writer.BaseStream.Position = position;
+        }
+
+        public static void SeekToEnd(this BinaryWriter writer)
+        {
+            writer.BaseStream.Position = writer.BaseStream.Length;
+        }
+
+        public static int GetLength(this BinaryReader reader)
+        {
+            return (int)reader.BaseStream.Length;
+        }
+
+        public static int GetLength(this BinaryWriter writer)
+        {
+            return (int)writer.BaseStream.Length;
+        }
+
+        public static void Skip(this BinaryReader reader, int size)
+        {
+            reader.BaseStream.Position += size;
         }
 
         public static void Align(this BinaryReader reader, int alignment)
