@@ -1,4 +1,6 @@
-﻿namespace SoundConverter
+﻿using TrRebootTools.SoundConverter;
+
+namespace TrRebootTools.SoundConverter
 {
     partial class MainForm
     {
@@ -32,18 +34,21 @@
             this._lstInputFiles = new System.Windows.Forms.ListBox();
             this._lblWavFiles = new System.Windows.Forms.Label();
             this._lblOutputFolder = new System.Windows.Forms.Label();
-            this._btnAddWavFile = new System.Windows.Forms.Button();
-            this._btnRemoveSelectedWavs = new System.Windows.Forms.Button();
-            this._btnClearWavFiles = new System.Windows.Forms.Button();
+            this._btnAddInputFiles = new System.Windows.Forms.Button();
+            this._btnRemoveSelectedInputFiles = new System.Windows.Forms.Button();
+            this._btnClearInputFiles = new System.Windows.Forms.Button();
             this._txtOutputFolder = new System.Windows.Forms.TextBox();
-            this._btnBrowseWemFolder = new System.Windows.Forms.Button();
+            this._btnBrowseOutputFolder = new System.Windows.Forms.Button();
             this._btnConvert = new System.Windows.Forms.Button();
             this._progressBar = new System.Windows.Forms.ProgressBar();
             this._btnCancel = new System.Windows.Forms.Button();
             this._pnlOptions = new System.Windows.Forms.Panel();
+            this._radShadow = new System.Windows.Forms.RadioButton();
+            this._radRise = new System.Windows.Forms.RadioButton();
+            this._radTr2013 = new System.Windows.Forms.RadioButton();
+            this.label1 = new System.Windows.Forms.Label();
             this._dlgSelectInputFiles = new System.Windows.Forms.OpenFileDialog();
             this._dlgSelectOutputFolder = new Ookii.Dialogs.WinForms.VistaFolderBrowserDialog();
-            this._dlgSelectWwiseConsole = new System.Windows.Forms.OpenFileDialog();
             this._pnlOptions.SuspendLayout();
             this.SuspendLayout();
             // 
@@ -56,22 +61,22 @@
             this._lstInputFiles.FormattingEnabled = true;
             this._lstInputFiles.IntegralHeight = false;
             this._lstInputFiles.ItemHeight = 12;
-            this._lstInputFiles.Location = new System.Drawing.Point(8, 27);
+            this._lstInputFiles.Location = new System.Drawing.Point(8, 65);
             this._lstInputFiles.Name = "_lstInputFiles";
             this._lstInputFiles.SelectionMode = System.Windows.Forms.SelectionMode.MultiSimple;
-            this._lstInputFiles.Size = new System.Drawing.Size(660, 223);
+            this._lstInputFiles.Size = new System.Drawing.Size(660, 185);
             this._lstInputFiles.TabIndex = 0;
-            this._lstInputFiles.DragDrop += new System.Windows.Forms.DragEventHandler(this._lstWavFiles_DragDrop);
-            this._lstInputFiles.DragEnter += new System.Windows.Forms.DragEventHandler(this._lstWavFiles_DragEnter);
+            this._lstInputFiles.DragDrop += new System.Windows.Forms.DragEventHandler(this._lstInputFiles_DragDrop);
+            this._lstInputFiles.DragEnter += new System.Windows.Forms.DragEventHandler(this._lstInputFiles_DragEnter);
             // 
             // _lblWavFiles
             // 
             this._lblWavFiles.AutoSize = true;
-            this._lblWavFiles.Location = new System.Drawing.Point(6, 10);
+            this._lblWavFiles.Location = new System.Drawing.Point(3, 50);
             this._lblWavFiles.Name = "_lblWavFiles";
-            this._lblWavFiles.Size = new System.Drawing.Size(84, 12);
+            this._lblWavFiles.Size = new System.Drawing.Size(58, 12);
             this._lblWavFiles.TabIndex = 1;
-            this._lblWavFiles.Text = "Input .wav files:";
+            this._lblWavFiles.Text = "Input files:";
             // 
             // _lblOutputFolder
             // 
@@ -79,42 +84,42 @@
             this._lblOutputFolder.AutoSize = true;
             this._lblOutputFolder.Location = new System.Drawing.Point(6, 320);
             this._lblOutputFolder.Name = "_lblOutputFolder";
-            this._lblOutputFolder.Size = new System.Drawing.Size(147, 12);
+            this._lblOutputFolder.Size = new System.Drawing.Size(74, 12);
             this._lblOutputFolder.TabIndex = 2;
-            this._lblOutputFolder.Text = "Output folder for .wem files:";
+            this._lblOutputFolder.Text = "Output folder:";
             // 
-            // _btnAddWavFile
+            // _btnAddInputFiles
             // 
-            this._btnAddWavFile.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
-            this._btnAddWavFile.Location = new System.Drawing.Point(8, 256);
-            this._btnAddWavFile.Name = "_btnAddWavFile";
-            this._btnAddWavFile.Size = new System.Drawing.Size(124, 36);
-            this._btnAddWavFile.TabIndex = 3;
-            this._btnAddWavFile.Text = "Add files...";
-            this._btnAddWavFile.UseVisualStyleBackColor = true;
-            this._btnAddWavFile.Click += new System.EventHandler(this._btnAddWavFile_Click);
+            this._btnAddInputFiles.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this._btnAddInputFiles.Location = new System.Drawing.Point(8, 256);
+            this._btnAddInputFiles.Name = "_btnAddInputFiles";
+            this._btnAddInputFiles.Size = new System.Drawing.Size(124, 36);
+            this._btnAddInputFiles.TabIndex = 3;
+            this._btnAddInputFiles.Text = "Add files...";
+            this._btnAddInputFiles.UseVisualStyleBackColor = true;
+            this._btnAddInputFiles.Click += new System.EventHandler(this._btnAddInputFiles_Click);
             // 
-            // _btnRemoveSelectedWavs
+            // _btnRemoveSelectedInputFiles
             // 
-            this._btnRemoveSelectedWavs.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this._btnRemoveSelectedWavs.Location = new System.Drawing.Point(391, 256);
-            this._btnRemoveSelectedWavs.Name = "_btnRemoveSelectedWavs";
-            this._btnRemoveSelectedWavs.Size = new System.Drawing.Size(151, 36);
-            this._btnRemoveSelectedWavs.TabIndex = 3;
-            this._btnRemoveSelectedWavs.Text = "Remove selected";
-            this._btnRemoveSelectedWavs.UseVisualStyleBackColor = true;
-            this._btnRemoveSelectedWavs.Click += new System.EventHandler(this._btnRemoveSelectedWavs_Click);
+            this._btnRemoveSelectedInputFiles.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
+            this._btnRemoveSelectedInputFiles.Location = new System.Drawing.Point(391, 256);
+            this._btnRemoveSelectedInputFiles.Name = "_btnRemoveSelectedInputFiles";
+            this._btnRemoveSelectedInputFiles.Size = new System.Drawing.Size(151, 36);
+            this._btnRemoveSelectedInputFiles.TabIndex = 3;
+            this._btnRemoveSelectedInputFiles.Text = "Remove selected";
+            this._btnRemoveSelectedInputFiles.UseVisualStyleBackColor = true;
+            this._btnRemoveSelectedInputFiles.Click += new System.EventHandler(this._btnRemoveSelectedInputFiles_Click);
             // 
-            // _btnClearWavFiles
+            // _btnClearInputFiles
             // 
-            this._btnClearWavFiles.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this._btnClearWavFiles.Location = new System.Drawing.Point(548, 256);
-            this._btnClearWavFiles.Name = "_btnClearWavFiles";
-            this._btnClearWavFiles.Size = new System.Drawing.Size(120, 36);
-            this._btnClearWavFiles.TabIndex = 3;
-            this._btnClearWavFiles.Text = "Clear";
-            this._btnClearWavFiles.UseVisualStyleBackColor = true;
-            this._btnClearWavFiles.Click += new System.EventHandler(this._btnClearWavFiles_Click);
+            this._btnClearInputFiles.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
+            this._btnClearInputFiles.Location = new System.Drawing.Point(548, 256);
+            this._btnClearInputFiles.Name = "_btnClearInputFiles";
+            this._btnClearInputFiles.Size = new System.Drawing.Size(120, 36);
+            this._btnClearInputFiles.TabIndex = 3;
+            this._btnClearInputFiles.Text = "Clear";
+            this._btnClearInputFiles.UseVisualStyleBackColor = true;
+            this._btnClearInputFiles.Click += new System.EventHandler(this._btnClearInputFiles_Click);
             // 
             // _txtOutputFolder
             // 
@@ -125,16 +130,16 @@
             this._txtOutputFolder.Size = new System.Drawing.Size(613, 19);
             this._txtOutputFolder.TabIndex = 4;
             // 
-            // _btnBrowseWemFolder
+            // _btnBrowseOutputFolder
             // 
-            this._btnBrowseWemFolder.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this._btnBrowseWemFolder.Location = new System.Drawing.Point(627, 334);
-            this._btnBrowseWemFolder.Name = "_btnBrowseWemFolder";
-            this._btnBrowseWemFolder.Size = new System.Drawing.Size(41, 21);
-            this._btnBrowseWemFolder.TabIndex = 5;
-            this._btnBrowseWemFolder.Text = "...";
-            this._btnBrowseWemFolder.UseVisualStyleBackColor = true;
-            this._btnBrowseWemFolder.Click += new System.EventHandler(this._btnBrowseWemFolder_Click);
+            this._btnBrowseOutputFolder.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
+            this._btnBrowseOutputFolder.Location = new System.Drawing.Point(627, 334);
+            this._btnBrowseOutputFolder.Name = "_btnBrowseOutputFolder";
+            this._btnBrowseOutputFolder.Size = new System.Drawing.Size(41, 20);
+            this._btnBrowseOutputFolder.TabIndex = 5;
+            this._btnBrowseOutputFolder.Text = "...";
+            this._btnBrowseOutputFolder.UseVisualStyleBackColor = true;
+            this._btnBrowseOutputFolder.Click += new System.EventHandler(this._btnBrowseOutputFolder_Click);
             // 
             // _btnConvert
             // 
@@ -175,12 +180,16 @@
             this._pnlOptions.Anchor = ((System.Windows.Forms.AnchorStyles)((((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Bottom) 
             | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this._pnlOptions.Controls.Add(this._btnBrowseWemFolder);
+            this._pnlOptions.Controls.Add(this._radShadow);
+            this._pnlOptions.Controls.Add(this._radRise);
+            this._pnlOptions.Controls.Add(this._radTr2013);
+            this._pnlOptions.Controls.Add(this._btnBrowseOutputFolder);
             this._pnlOptions.Controls.Add(this._txtOutputFolder);
-            this._pnlOptions.Controls.Add(this._btnClearWavFiles);
-            this._pnlOptions.Controls.Add(this._btnRemoveSelectedWavs);
-            this._pnlOptions.Controls.Add(this._btnAddWavFile);
+            this._pnlOptions.Controls.Add(this._btnClearInputFiles);
+            this._pnlOptions.Controls.Add(this._btnRemoveSelectedInputFiles);
+            this._pnlOptions.Controls.Add(this._btnAddInputFiles);
             this._pnlOptions.Controls.Add(this._lblOutputFolder);
+            this._pnlOptions.Controls.Add(this.label1);
             this._pnlOptions.Controls.Add(this._lblWavFiles);
             this._pnlOptions.Controls.Add(this._lstInputFiles);
             this._pnlOptions.Location = new System.Drawing.Point(12, 6);
@@ -188,19 +197,55 @@
             this._pnlOptions.Size = new System.Drawing.Size(670, 379);
             this._pnlOptions.TabIndex = 9;
             // 
+            // _radShadow
+            // 
+            this._radShadow.AutoSize = true;
+            this._radShadow.Location = new System.Drawing.Point(192, 18);
+            this._radShadow.Name = "_radShadow";
+            this._radShadow.Size = new System.Drawing.Size(60, 16);
+            this._radShadow.TabIndex = 6;
+            this._radShadow.Text = "SOTTR";
+            this._radShadow.UseVisualStyleBackColor = true;
+            // 
+            // _radRise
+            // 
+            this._radRise.AutoSize = true;
+            this._radRise.Location = new System.Drawing.Point(97, 18);
+            this._radRise.Name = "_radRise";
+            this._radRise.Size = new System.Drawing.Size(61, 16);
+            this._radRise.TabIndex = 6;
+            this._radRise.Text = "ROTTR";
+            this._radRise.UseVisualStyleBackColor = true;
+            // 
+            // _radTr2013
+            // 
+            this._radTr2013.AutoSize = true;
+            this._radTr2013.Checked = true;
+            this._radTr2013.Location = new System.Drawing.Point(8, 18);
+            this._radTr2013.Name = "_radTr2013";
+            this._radTr2013.Size = new System.Drawing.Size(62, 16);
+            this._radTr2013.TabIndex = 6;
+            this._radTr2013.TabStop = true;
+            this._radTr2013.Text = "TR2013";
+            this._radTr2013.UseVisualStyleBackColor = true;
+            // 
+            // label1
+            // 
+            this.label1.AutoSize = true;
+            this.label1.Location = new System.Drawing.Point(3, 3);
+            this.label1.Name = "label1";
+            this.label1.Size = new System.Drawing.Size(36, 12);
+            this.label1.TabIndex = 1;
+            this.label1.Text = "Game:";
+            // 
             // _dlgSelectInputFiles
             // 
-            this._dlgSelectInputFiles.Filter = "Wave files|*.wav";
             this._dlgSelectInputFiles.Multiselect = true;
-            // 
-            // _dlgSelectWwiseConsole
-            // 
-            this._dlgSelectWwiseConsole.Filter = "WwiseConsole.exe|WwiseConsole.exe";
             // 
             // MainForm
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 12F);
-            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.AutoScaleDimensions = new System.Drawing.SizeF(96F, 96F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Dpi;
             this.ClientSize = new System.Drawing.Size(694, 444);
             this.Controls.Add(this._btnConvert);
             this.Controls.Add(this._pnlOptions);
@@ -209,7 +254,7 @@
             this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
             this.MinimumSize = new System.Drawing.Size(590, 360);
             this.Name = "MainForm";
-            this.Text = "Wwise Sound Conversion Helper";
+            this.Text = "TR Reboot Sound Converter";
             this.FormClosing += new System.Windows.Forms.FormClosingEventHandler(this.MainForm_FormClosing);
             this.Load += new System.EventHandler(this.MainForm_Load);
             this._pnlOptions.ResumeLayout(false);
@@ -223,18 +268,21 @@
         private System.Windows.Forms.ListBox _lstInputFiles;
         private System.Windows.Forms.Label _lblWavFiles;
         private System.Windows.Forms.Label _lblOutputFolder;
-        private System.Windows.Forms.Button _btnAddWavFile;
-        private System.Windows.Forms.Button _btnRemoveSelectedWavs;
-        private System.Windows.Forms.Button _btnClearWavFiles;
+        private System.Windows.Forms.Button _btnAddInputFiles;
+        private System.Windows.Forms.Button _btnRemoveSelectedInputFiles;
+        private System.Windows.Forms.Button _btnClearInputFiles;
         private System.Windows.Forms.TextBox _txtOutputFolder;
-        private System.Windows.Forms.Button _btnBrowseWemFolder;
+        private System.Windows.Forms.Button _btnBrowseOutputFolder;
         private System.Windows.Forms.Button _btnConvert;
         private System.Windows.Forms.ProgressBar _progressBar;
         private System.Windows.Forms.Button _btnCancel;
         private System.Windows.Forms.Panel _pnlOptions;
         private System.Windows.Forms.OpenFileDialog _dlgSelectInputFiles;
         private Ookii.Dialogs.WinForms.VistaFolderBrowserDialog _dlgSelectOutputFolder;
-        private System.Windows.Forms.OpenFileDialog _dlgSelectWwiseConsole;
+        private System.Windows.Forms.RadioButton _radShadow;
+        private System.Windows.Forms.RadioButton _radRise;
+        private System.Windows.Forms.RadioButton _radTr2013;
+        private System.Windows.Forms.Label label1;
     }
 }
 
