@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using TrRebootTools.SoundConverter;
 
-namespace SoundConverter
+namespace TrRebootTools.SoundConverter.Properties
 {
     internal static class Program
     {
