@@ -3,7 +3,7 @@ bl_info = {
     "description": "Import/export files for the Tomb Raider Reboot games",
     "author": "arc_",
     "blender": (4, 0, 0),
-    "version": (1, 7, 0),
+    "version": (1, 7, 1),
     "location": "File > Import-Export",
     "support": "COMMUNITY",
     "category": "Import-Export"
