@@ -3,10 +3,9 @@ import math
 from typing import Any, ClassVar, Literal, NamedTuple, cast
 from mathutils import Matrix, Quaternion, Vector
 from io_scene_tr_reboot.BlenderHelper import BlenderHelper
-from io_scene_tr_reboot.BlenderNaming import BlenderBoneIdSet, BlenderNaming
+from io_scene_tr_reboot.BlenderNaming import BlenderNaming
 from io_scene_tr_reboot.properties.BoneProperties import BoneProperties
 from io_scene_tr_reboot.properties.ObjectProperties import ObjectSkeletonProperties
-from io_scene_tr_reboot.tr.Bone import IBone
 from io_scene_tr_reboot.tr.BoneConstraint import BoneConstraintType, IBoneConstraint, IBoneConstraint_WeightedPosition, IBoneConstraint_WeightedRotation, IBoneConstraint_LookAt
 from io_scene_tr_reboot.tr.Collection import Collection
 from io_scene_tr_reboot.tr.ResourceKey import ResourceKey
@@ -20,11 +19,6 @@ class _BoneTransform(NamedTuple):
     tail: Vector
     orientation: Quaternion
     is_flipped: bool
-
-class _ConstrainedBoneSet(NamedTuple):
-    tr_bone: IBone
-    bl_constrained_pose_bone: bpy.types.PoseBone
-    bl_helper_pose_bone: bpy.types.PoseBone | None
 
 class SkeletonImporter(SlotsBase):
     scale_factor: float
@@ -56,7 +50,7 @@ class SkeletonImporter(SlotsBase):
                 bl_armature_obj = BlenderHelper.create_object(bl_armature)
 
                 bone_transforms = self.calc_bone_transforms(tr_skeleton)
-                self.create_bones(bl_armature, tr_skeleton, bone_transforms)
+                self.create_bones(bl_armature_obj, tr_skeleton, bone_transforms)
                 self.assign_counterparts(bl_armature_obj, tr_skeleton)
                 self.create_constraints(bl_armature_obj, tr_skeleton, bone_transforms)
             else:
@@ -133,7 +127,8 @@ class SkeletonImporter(SlotsBase):
 
         return bone_flip_states
 
-    def create_bones(self, bl_armature: bpy.types.Armature, tr_skeleton: ISkeleton, bone_transforms: list[_BoneTransform]) -> None:
+    def create_bones(self, bl_armature_obj: bpy.types.Object, tr_skeleton: ISkeleton, bone_transforms: list[_BoneTransform]) -> None:
+        bl_armature = cast(bpy.types.Armature, bl_armature_obj.data)
         with BlenderHelper.enter_edit_mode():
             for i, tr_bone in enumerate(tr_skeleton.bones):
                 bl_bone = bl_armature.edit_bones.new(BlenderNaming.make_bone_name(None, tr_bone.global_id, i))
@@ -150,6 +145,8 @@ class SkeletonImporter(SlotsBase):
         for i, tr_bone in enumerate(tr_skeleton.bones):
             bl_bone = bl_armature.bones[BlenderNaming.make_bone_name(None, tr_bone.global_id, i)]
             BoneProperties.get_instance(bl_bone).is_flipped = bone_transforms[i].is_flipped
+            if len(tr_bone.constraints) > 0:
+                BlenderHelper.add_bone_to_group(bl_armature_obj, bl_bone, BlenderNaming.constrained_bone_group_name, BlenderNaming.constrained_bone_palette_name)
 
     def assign_counterparts(self, bl_armature_obj: bpy.types.Object, tr_skeleton: ISkeleton) -> None:
         bl_armature = cast(bpy.types.Armature, bl_armature_obj.data)
@@ -163,74 +160,32 @@ class SkeletonImporter(SlotsBase):
 
     def create_constraints(self, bl_armature_obj: bpy.types.Object, tr_skeleton: ISkeleton, bone_transforms: list[_BoneTransform]) -> None:
         if bl_armature_obj.pose is None:
-            raise Exception("Armature has no pose")
+            return
 
-        constrained_bone_sets = self.create_helper_bones(bl_armature_obj, tr_skeleton)
-        for constrained_bone_set in constrained_bone_sets:
-            for tr_constraint in constrained_bone_set.tr_bone.constraints:
-                self.create_constraint(bl_armature_obj, bone_transforms, constrained_bone_set, tr_constraint)
-
-        # Hack to make the now-constrained bones settle into place
-        with BlenderHelper.enter_edit_mode(bl_armature_obj):
-            pass
-
-    def create_helper_bones(self, bl_armature_obj: bpy.types.Object, tr_skeleton: ISkeleton) -> list[_ConstrainedBoneSet]:
-        if bl_armature_obj.pose is None:
-            return []
-
-        bl_armature = cast(bpy.types.Armature, bl_armature_obj.data)
-        constrained_bone_id_sets: list[BlenderBoneIdSet] = []
-        with BlenderHelper.enter_edit_mode(bl_armature_obj):
-            for i, tr_bone in enumerate(tr_skeleton.bones):
-                if len(tr_bone.constraints) == 0 or tr_bone.global_id is None:
-                    continue
-
-                constrained_bone_id_sets.append(BlenderBoneIdSet(None, tr_bone.global_id, i))
-
-                bl_helper_bone = bl_armature.edit_bones.new(BlenderNaming.make_helper_bone_name(tr_bone.global_id))
-                bl_helper_bone.tail = (0, 1, 0)
-
-        constrained_bone_sets: list[_ConstrainedBoneSet] = []
-        for constrained_bone_id_set in constrained_bone_id_sets:
-            if constrained_bone_id_set.global_id is None or constrained_bone_id_set.local_id is None:
-                raise Exception()
-
-            tr_bone = tr_skeleton.bones[constrained_bone_id_set.local_id]
-
-            constrained_bone_name = BlenderNaming.make_bone_name(constrained_bone_id_set)
-            bl_constrained_bone = bl_armature.bones[constrained_bone_name]
-            bl_constrained_pose_bone = bl_armature_obj.pose.bones[constrained_bone_name]
-            BlenderHelper.add_bone_to_group(bl_armature_obj, bl_constrained_bone, BlenderNaming.constrained_bone_group_name, BlenderNaming.constrained_bone_palette_name)
-
-            helper_bone_name = BlenderNaming.make_helper_bone_name(constrained_bone_id_set.global_id)
-            bl_helper_bone = bl_armature.bones[helper_bone_name]
-            bl_helper_pose_bone = bl_armature_obj.pose.bones[helper_bone_name]
-            BlenderHelper.add_bone_to_group(bl_armature_obj, bl_helper_bone, BlenderNaming.helper_bone_group_name, BlenderNaming.helper_bone_palette_name, False)
-
-            constrained_bone_sets.append(_ConstrainedBoneSet(tr_bone, bl_constrained_pose_bone, bl_helper_pose_bone))
-
-        return constrained_bone_sets
+        for i, tr_bone in enumerate(tr_skeleton.bones):
+            bl_pose_bone = bl_armature_obj.pose.bones[BlenderNaming.make_bone_name(None, tr_bone.global_id, i)]
+            for tr_constraint in tr_bone.constraints:
+                self.create_constraint(bl_armature_obj, bone_transforms, bl_pose_bone, tr_constraint)
 
     def create_constraint(
         self,
         bl_armature_obj: bpy.types.Object,
         bone_transforms: list[_BoneTransform],
-        constrained_bone_set: _ConstrainedBoneSet,
+        bl_pose_bone: bpy.types.PoseBone,
         tr_constraint: IBoneConstraint
     ) -> None:
-        bl_bone = constrained_bone_set.bl_constrained_pose_bone.bone
-        constraint_properties = BoneProperties.get_instance(bl_bone).constraints.add()
+        constraint_properties = BoneProperties.get_instance(bl_pose_bone.bone).constraints.add()
         constraint_properties.data = tr_constraint.serialize()
 
         match tr_constraint.type:
             case BoneConstraintType.LOOK_AT:
-                self.create_look_at_constraint(bl_armature_obj, bone_transforms, constrained_bone_set, cast(IBoneConstraint_LookAt, tr_constraint))
+                self.create_look_at_constraint(bl_armature_obj, bone_transforms, bl_pose_bone, cast(IBoneConstraint_LookAt, tr_constraint))
 
             case BoneConstraintType.WEIGHTED_POSITION:
-                self.create_weighted_position_constraint(bl_armature_obj, bone_transforms, constrained_bone_set, cast(IBoneConstraint_WeightedPosition, tr_constraint))
+                self.create_weighted_position_constraint(bl_armature_obj, bone_transforms, bl_pose_bone, cast(IBoneConstraint_WeightedPosition, tr_constraint))
 
             case BoneConstraintType.WEIGHTED_ROTATION:
-                self.create_weighted_rotation_constraint(bl_armature_obj, bone_transforms, constrained_bone_set, cast(IBoneConstraint_WeightedRotation, tr_constraint))
+                self.create_weighted_rotation_constraint(bl_armature_obj, bone_transforms, bl_pose_bone, cast(IBoneConstraint_WeightedRotation, tr_constraint))
 
             case _:
                 pass
@@ -239,17 +194,14 @@ class SkeletonImporter(SlotsBase):
         self,
         bl_armature_obj: bpy.types.Object,
         bone_transforms: list[_BoneTransform],
-        constrained_bone_set: _ConstrainedBoneSet,
+        bl_pose_bone: bpy.types.PoseBone,
         tr_constraint: IBoneConstraint_LookAt
     ) -> None:
-        (tr_bone, bl_constrained_pose_bone, bl_helper_pose_bone) = constrained_bone_set
-        if tr_bone.parent_id < 0 or bl_helper_pose_bone is None:
-            return
-
-        bl_curves = cast(list[bpy.types.FCurve], bl_helper_pose_bone.driver_add("rotation_quaternion"))
+        bl_curves = cast(list[bpy.types.FCurve], bl_pose_bone.driver_add("rotation_quaternion"))
         for elem_idx, bl_curve in enumerate(bl_curves):
             bl_driver = cast(bpy.types.Driver, bl_curve.driver)
-            armature_rotation = self.make_driver_expr_for_obj_attr(bl_driver, bl_armature_obj, None, "rotation_quaternion")
+            bl_driver.use_self = True
+            #armature_rotation = self.make_driver_expr_for_obj_attr(bl_driver, bl_armature_obj, None, "rotation_quaternion")
             looker_position   = self.make_driver_expr_for_bone_attr(bl_driver, bl_armature_obj, bone_transforms, tr_constraint.target_bone_local_id, "location")
             look_at_positions = self.make_driver_expr_for_bones_attr(bl_driver, bl_armature_obj, bone_transforms, tr_constraint.source_bone_local_ids, "location")
             look_at_weights   = self.float_list_to_string(tr_constraint.source_bone_weights)
@@ -279,48 +231,33 @@ class SkeletonImporter(SlotsBase):
 
             bone_local_tangent_expr = self.float_tuple_to_string(bone_local_tangent)
             bone_local_normal_expr  = self.float_tuple_to_string(bone_local_normal)
-            bl_driver.expression = f"tr_look_at({armature_rotation},{looker_position},{look_at_positions},{look_at_weights},{pole_dir_expr},{bone_local_tangent_expr},{bone_local_normal_expr})[{elem_idx}]"
-
-        bl_constraint = cast(bpy.types.CopyRotationConstraint, bl_constrained_pose_bone.constraints.new("COPY_ROTATION"))
-        bl_constraint.target = bl_armature_obj
-        bl_constraint.subtarget = bl_helper_pose_bone.name
+            bl_driver.expression = f"tr_look_at(self,{looker_position},{look_at_positions},{look_at_weights},{pole_dir_expr},{bone_local_tangent_expr},{bone_local_normal_expr})[{elem_idx}]"
 
     def create_weighted_position_constraint(
         self,
         bl_armature_obj: bpy.types.Object,
         bone_transforms: list[_BoneTransform],
-        constrained_bone_set: _ConstrainedBoneSet,
+        bl_pose_bone: bpy.types.PoseBone,
         tr_constraint: IBoneConstraint_WeightedPosition
     ) -> None:
-        (_, bl_constrained_pose_bone, bl_helper_pose_bone) = constrained_bone_set
-        if bl_helper_pose_bone is None:
-            return
-
         self.create_weighted_attr_drivers(
             bl_armature_obj,
             bone_transforms,
-            bl_helper_pose_bone,
+            bl_pose_bone,
             "tr_weighted_pos",
             tr_constraint.source_bone_local_ids,
             tr_constraint.source_bone_weights,
             tr_constraint.offset,
             "location"
         )
-        bl_constraint = cast(bpy.types.CopyLocationConstraint, bl_constrained_pose_bone.constraints.new("COPY_LOCATION"))
-        bl_constraint.target = bl_armature_obj
-        bl_constraint.subtarget = bl_helper_pose_bone.name
 
     def create_weighted_rotation_constraint(
         self,
         bl_armature_obj: bpy.types.Object,
         bone_transforms: list[_BoneTransform],
-        constrained_bone_set: _ConstrainedBoneSet,
+        bl_pose_bone: bpy.types.PoseBone,
         tr_constraint: IBoneConstraint_WeightedRotation
     ) -> None:
-        (_, bl_constrained_pose_bone, bl_helper_pose_bone) = constrained_bone_set
-        if bl_helper_pose_bone is None:
-            return
-
         offset = tr_constraint.offset
         if bone_transforms[tr_constraint.target_bone_local_id].is_flipped:
             offset = offset @ SkeletonImporter.tr_flip_quat
@@ -328,38 +265,36 @@ class SkeletonImporter(SlotsBase):
         self.create_weighted_attr_drivers(
             bl_armature_obj,
             bone_transforms,
-            bl_helper_pose_bone,
+            bl_pose_bone,
             "tr_weighted_rot",
             tr_constraint.source_bone_local_ids,
             tr_constraint.source_bone_weights,
             offset,
             "rotation_quaternion"
         )
-        bl_constraint = cast(bpy.types.CopyRotationConstraint, bl_constrained_pose_bone.constraints.new("COPY_ROTATION"))
-        bl_constraint.target = bl_armature_obj
-        bl_constraint.subtarget = bl_helper_pose_bone.name
 
     def create_weighted_attr_drivers(
         self,
         bl_armature_obj: bpy.types.Object,
         bone_transforms: list[_BoneTransform],
-        bl_helper_pose_bone: bpy.types.PoseBone,
+        bl_pose_bone: bpy.types.PoseBone,
         driver_func_name: str,
         source_bone_local_ids: list[int],
         source_bone_weights: list[float],
         offset: Vector | Quaternion,
         attr_name: Literal["location"] | Literal["rotation_quaternion"]
     ) -> None:
-        bl_curves = cast(list[bpy.types.FCurve], bl_helper_pose_bone.driver_add(attr_name))
+        bl_curves = cast(list[bpy.types.FCurve], bl_pose_bone.driver_add(attr_name))
         for elem_idx, bl_curve in enumerate(bl_curves):
             bl_driver = cast(bpy.types.Driver, bl_curve.driver)
-            armature_position = self.make_driver_expr_for_obj_attr(bl_driver, bl_armature_obj, None, "location")
-            armature_rotation = self.make_driver_expr_for_obj_attr(bl_driver, bl_armature_obj, None, "rotation_quaternion")
+            bl_driver.use_self = True
+            #armature_position = self.make_driver_expr_for_obj_attr(bl_driver, bl_armature_obj, None, "location")
+            #armature_rotation = self.make_driver_expr_for_obj_attr(bl_driver, bl_armature_obj, None, "rotation_quaternion")
             bone_attrs        = self.make_driver_expr_for_bones_attr(bl_driver, bl_armature_obj, bone_transforms, source_bone_local_ids, attr_name)
             bone_flip_flags   = self.bone_flip_flags_to_string(bone_transforms, source_bone_local_ids)
             weights_expr      = self.float_list_to_string(source_bone_weights)
             offset_expr       = self.float_tuple_to_string(offset)
-            bl_driver.expression = f"{driver_func_name}({armature_position},{armature_rotation},{bone_attrs},{bone_flip_flags},{weights_expr},{offset_expr})[{elem_idx}]"
+            bl_driver.expression = f"{driver_func_name}(self,{bone_attrs},{bone_flip_flags},{weights_expr},{offset_expr})[{elem_idx}]"
 
     def make_driver_expr_for_bones_attr(
         self,
